Refactor to use only a single loader function
Rather than make a separate loader function for each extension, use just
one loader function and dispatch each extension individually.

Note that this means that if two or more loaders are used *and* both of
those loaders define a script with the same name (which is a bad idea),
it is non-deterministic which one will be found first since Lua does not
maintain key ordering in tables. But I do not consider this a bug since
two languages should not both have a script of the same name.